% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\docType{class}
\name{EnvimRed-class}
\alias{EnvimRed-class}
\alias{EnvimRed}
\title{myRefClass}
\description{
myRefClass
}
\section{Slots}{

\describe{
\item{\code{cropa}}{A RasterBrinck}

\item{\code{m.env}}{A Matrix}

\item{\code{project}}{A List}
}}

