% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytotoma.R
\docType{data}
\name{phytotoma}
\alias{phytotoma}
\title{Peruvian Plantcutter Occurrence Data}
\format{A data frame with 50 rows and 3 columns}
\usage{
data(phytotoma)
}
\description{
A dataset containing occurrence record locations of Peruvian plantcutter
 (\emph{Phyotoma raimondii}), endemic species of Dry forest
 in northern Peru. These records have been examined to check for accurate
 species identification.
}
\details{
\itemize{
  \item \code{sp.name} species name
  \item \code{x} Longitude
  \item \code{y} Latitude
}
}
\keyword{datasets}
