% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.env.R
\name{reduce.env}
\alias{reduce.env}
\title{Reduce environmental data}
\usage{
reduce.env(env, transfer = NULL, var.rm, mask)
}
\arguments{
\item{env}{RasterStack* objet.}

\item{transfer}{List of rasterstack object}

\item{var.rm}{A data.frame of occurrence records.
It must include two column based on latitude and longitude.}

\item{mask}{Croped mask, must be shapefile (.shp), readOGR.}
}
\value{
\code{@crop} RasterStack* Objet

 \code{@m.env} data.frame of environmental values to occurrence localities.
}
\description{
This function allows reduce environmental data clipping by mask or buffer area
}
\details{
Reduce the correlation among predicted variables either buffer zone, or
clipping mask.

Provide reduce objet with cut predicted variables and data.frame for
the values of each point of occurrence of them.
}
\examples{

# Phytotoma ocurrence data
data(phytotoma)


# Complement
library(dismo)
predictor <- stack(list.files(path=paste(system.file(package="dismo"),'/ex', sep=''),
 pattern='grd', full.names=TRUE ))

 maskM <- stim.M(phytotoma[,2:3], 131)

vrn <- extract(predictor, phytotoma[,2:3])

# Correlogram
rd <- c('bio_1','bio_12','biome')
cor.show(vrn, rm=TRUE, var.rm=rd)

# Reduce spatial-data
reduce_cut <- reduce.env(env = predictor, var.rm=rd, mask=maskM)

# Plot reduce_cut
plot(reduce_cut@cropa$bio16)

# Add points
points(phytotoma[,2:3], pch=16,col='blue')

# summary
reduce_cut


}
\seealso{
\code{\link{cor.show}}
}
