% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{stability}
\alias{stability}
\title{Stability of ecological niche models}
\usage{
stability(current = NULL, future = NULL, thr.value = NULL,
  charname = NULL)
}
\arguments{
\item{current}{Raster* objet of present distribution.
Raster has continue values.}

\item{future}{RasterStack* object of future distributions.
Must have three models with continue values.}

\item{thr.value}{Cut value (0 — 1) of threshold in order to
species distribution.}

\item{charname}{\code{character-Class}, name of the species.
(\code{default = NULL}).}
}
\value{
An object of class 'StabEcodist'

\bold{Method S4 print}

Return table with these features: \code{Models} and \code{nPixel}
(frequency of pixel with that feature).

\enumerate{

\item{\bold{Lost area}

{code \code{Models}:

 \code{code 0} mention the area with species absence.

 \code{code 14} mention the lost area current in the future projections}
}

\item{\bold{Gain area}

{code \code{Models}:

 \code{code 1} mention the gain area acording with one model.

 \code{code 2} acording with two models.

 \code{code 3} acording with three models. }
}

\item{\bold{Stable}

{code \code{Models}:

 \code{code 15} mention the stable area with one model.

 \code{code 16} mention the stable area with two models

 \code{code 17} mention the stable area with three models.}
}

}

\bold{Method S4 plot}

Return maps of stability niche
}
\description{
Returns stability niche based on species distribution models and their projections.
}
\references{
Peterson et al., (2017) Influences of climate change on the potential
distribution of Lutzomyia longipalpis sensu lato (Psychodidae: Phlebotominae).
 International Journal for Parasitology. 47(10–11):667–74.
}
