% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinOcc.R
\name{ThinOcc}
\alias{ThinOcc}
\title{Thin Spatial Occurrence}
\usage{
ThinOcc(
  df,
  NullRaster,
  SpeciesName = NULL,
  ThinRange = 1,
  Polygon = NULL,
  methods = c("spatial", "polygons"),
  save = TRUE,
  ...
)
}
\arguments{
\item{df}{dataframe-class. Occurrence data e.g. long,lat.}

\item{NullRaster}{raster-class.}

\item{SpeciesName}{character-class.}

\item{ThinRange}{numeric-class. Radio of filter species occurrence.}

\item{Polygon}{spatial-class.}

\item{methods}{Choose .}

\item{save}{logical.}

\item{...}{no implemented. Use \code{trace=TRUE} to show vif running. \code{default=FALSE}.}
}
\description{
Returns a data frame of occurrence dataset.
}
\details{
Thin ocurrence data using three approaches Polygons, Buffer zone, and Biogeographical regions.
}
\seealso{
\code{\link{reduce.env}}
}
