% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\docType{class}
\name{VarSelection-class}
\alias{VarSelection-class}
\title{VarSelection}
\description{
VarSelection
}
\section{Slots}{

\describe{
\item{\code{variables}}{A RasterBrinck}

\item{\code{VarsSets}}{A Matrix}
}}

