% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varSelection.R
\name{varSelection}
\alias{varSelection}
\title{Variable Selection}
\usage{
varSelection(
  df,
  VarMethod = NULL,
  vifs = NULL,
  cutoff = NULL,
  removeVariable = NULL,
  keep_var = FALSE,
  ...
)
}
\arguments{
\item{df}{dataframe-class.}

\item{VarMethod}{Method used to variable selection. \code{vif} uses variance inflaction factor 
based on \code{car} package. \code{correlation} uses pearson correlation to 
select random variable using cutoff parameter. \code{both} use a two approaches describe above.}

\item{vifs}{numeric vector. vif uses numeric vector to return variables selection.}

\item{cutoff}{numeric vector. Person correlation coef. used like threshold.}

\item{removeVariable}{Variables choosed to remove before running varSelection.}

\item{keep_var}{logical. If \code{TRUE}, all variables will be save in \code{Variables_Selected} folder.
(\code{default = FALSE}).}

\item{...}{no implemented. Use \code{trace=TRUE} to show vif running. \code{default=FALSE}.}
}
\value{
Correlogram plot
}
\description{
Returns plot with correlation values among predicted variables.
}
\details{
Variable selection using two approaches: vif and person correlation.
}
\examples{

DFset <- data.frame(
var1 = c(1,2,3,2,3,2,3,2,3,2,3,4,3,4,4),
var2 = c(2,3,6,4,5,4,5,6,4,2,3,4,3,6,1),
var3 = c(5,4,3,4,5,4,5,6,5,6,2,3,4,5,3),
var4 = c(5,4,3,1,2,3,5,3,3,3,2,4,3,3,4),
var5 = c(5,2,5,1,1,2,4,1,2,3,3,4,1,2,5),
var6 = c(10,12,23,34,23,34,23,12,9,23,12,34,12,23,9))

# Pearson correlation method
VarsKeep1 <- varSelection(DFset, VarMethod = 'correlation',  
                       cutoff=0.75)
                       
# vif correlation method
VarsKeep2 <- varSelection(DFset, VarMethod = 'vif', 
                          vifs=c(2,5))
                          
# both aprroaches: vif and correlation
VarsKeep3 <- varSelection(DFset, VarMethod = 'both', 
                          vifs=c(2,5), cutoff = c(0.2,0.5))

# both aprroaches: vif and correlation. Removing some variables and keep vars
VarsKeep4 <- varSelection(DFset, VarMethod = 'both', 
                          vifs=c(2,5), cutoff = c(0.2,0.5),
                          removeVariable=c('var1','var2'),
                          keep_var = TRUE)


}
\seealso{
\code{\link{reduce.env}}
}
