% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.show.R
\name{cor.show}
\alias{cor.show}
\title{Correlogram plot}
\usage{
cor.show(r, rm = FALSE, var.rm)
}
\arguments{
\item{r}{EnvimRed-class.}

\item{rm}{logical. If \code{TRUE}, allows remove some
variables from imput data set. (\code{default = FALSE})}

\item{var.rm}{variables names of imput data set. Using
\code{colnames(RasterStick)}, Where \code{RasterStick} is RasterStack* object.}
}
\value{
Correlogram plot
}
\description{
Returns plot with correlation values among predicted variables.
}
\details{
Correlation matrix based on pearson.
}
\seealso{
\code{\link{reduce.env}}
}
