% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varSelection.R
\name{varSelection}
\alias{varSelection}
\title{Variable Selection}
\usage{
varSelection(
  df,
  VarMethod = NULL,
  vifs = NULL,
  cutoff = NULL,
  removeVariable = NULL,
  keep_var = FALSE,
  ...
)
}
\arguments{
\item{df}{dataframe-class.}

\item{VarMethod}{Method used to variable selection. \code{vif} uses variance inflaction factor 
based on \code{car} package. \code{correlation} uses pearson correlation to 
select random variable using cutoff parameter. \code{both} use a two approaches describe above.}

\item{vifs}{numeric vector. vif uses numeric vector to return variables selection.}

\item{cutoff}{numeric vector. Person correlation coef. used like threshold.}

\item{removeVariable}{Variables choosed to remove before running varSelection.}

\item{keep_var}{logical. If \code{TRUE}, all variables will be save in \code{Variables_Selected} folder.
(\code{default = FALSE}).}

\item{...}{no implemented. Use \code{trace=TRUE} to show vif running. \code{default=FALSE}.}
}
\value{
Correlogram plot
}
\description{
Returns plot with correlation values among predicted variables.
}
\details{
Variable selection using two approaches: vif and person correlation.
}
\seealso{
\code{\link{reduce.env}}
}
